requirement.txt
pdfplumber
Pillow
pdf2image
pytesseract
pandas
numpy
tqdm
